+++ geompp/src/calc_utils3d.cpp
@@ -1,5 +1,6 @@
 #include "calc_utils3d.hpp"
 
+#include "calc_utils2d.hpp"
 #include "utils.hpp"
 #include "vector3d.hpp"
 
@@ -107,6 +108,92 @@ std::optional<Point3D> intersection_line_to_line(Point3D const& L1_P0, Point3D c
 
   sc = tc = std::numeric_limits<double>::quiet_NaN();
   return std::nullopt;
+}
+
+std::vector<std::size_t> convex_hull_indices(std::vector<Point3D> const& points, std::optional<Vector3D> normal) {
+  if (points.size() < 3) {
+    throw std::invalid_argument("less than 3 points");
+  }
+
+  // TODO: can this block become a function that makes sense to use ?
+  //       Vector3D make_normal_or_throw(std::vector<Point3D> const& cloud_of_coplanar_points);
+  if (!normal) {
+    auto no_col_pts = remove_collinear(points);
+    if (no_col_pts.size() < 3) {
+      throw std::invalid_argument("less than 3 non collinear points");
+    }
+
+    Vector3D calc_normal = (no_col_pts[1] - no_col_pts[0]).Cross(no_col_pts[2] - no_col_pts[0]);
+
+    std::size_t n = no_col_pts.size();
+    if (n > 3) {
+      for (std::size_t i = 3; i < n + 1; ++i) {
+        auto p0 = no_col_pts[(i - 2) % n];
+        auto p1 = no_col_pts[(i - 1) % n];
+        auto p2 = no_col_pts[i % n];
+
+        // performance note: here we try to assess if all points are on the same plane - that is - all normals are equal
+        //  I would have used .Normalize() and compare pairs of normals but that would require 3 sqrt() calls each time
+        //  It is faster to use a cross product of the two calculated normals and check if they cross to a null vector
+        //  (if they don't they aren't on the same plane)
+        auto temp_norm = (p1 - p0).Cross(p2 - p0);
+        if (!calc_normal.Cross(temp_norm).AlmostEquals(Vector3D{0, 0, 0})) {
+          throw std::invalid_argument("points are not co-planar");
+        }
+      }
+    }
+
+    if (calc_normal.AlmostEquals(Vector3D{0, 0, 0})) {
+      throw std::logic_error("failed to calculate points normal");
+    }
+
+    normal = calc_normal;
+  }
+
+  auto dax = normal->DominantAxis();
+
+  switch (dax) {
+    case Axis::X:
+      // clang-format off
+      return convex_hull_generic_impl_2D(
+          points.size(),
+          [&points](size_t i) { return points[i].y(); },
+          [&points](size_t i) { return points[i].z(); },
+          [&points](size_t o, size_t a, size_t b) {
+              return compare((points[a].y() - points[o].y()) * (points[b].z() - points[o].z()) -
+                             (points[a].z() - points[o].z()) * (points[b].y() - points[o].y()), 0) > 0;
+          }
+      );
+      // clang-format on
+    case Axis::Y:
+      // clang-format off
+      return convex_hull_generic_impl_2D(
+          points.size(),
+          [&points](size_t i) { return points[i].z(); },
+          [&points](size_t i) { return points[i].x(); },
+          [&points](size_t o, size_t a, size_t b) {
+              return compare((points[a].z() - points[o].z()) * (points[b].x() - points[o].x()) -
+                             (points[a].x() - points[o].x()) * (points[b].z() - points[o].z()), 0) > 0;
+          }
+      );
+      // clang-format on
+
+    case Axis::Z:
+      // clang-format off
+      return convex_hull_generic_impl_2D(
+          points.size(),
+          [&points](size_t i) { return points[i].x(); },
+          [&points](size_t i) { return points[i].y(); },
+          [&points](size_t o, size_t a, size_t b) {
+              return compare((points[a].x() - points[o].x()) * (points[b].y() - points[o].y()) -
+                             (points[a].y() - points[o].y()) * (points[b].x() - points[o].x()), 0) > 0;
+          }
+      );
+      // clang-format on
+
+    default:
+      throw std::logic_error("unexpected dominant axis");
+  }
 }
 
 }  // namespace geompp